Bugtracker - fix comments triggering "opened"
